perf: 修改 model 继承
